+++ options/defaults.lua
@@ -699,6 +699,9 @@ SAO.defaults = {
                     [78]  = true, -- Heroic Strike
                     [845] = true, -- Cleave
                 },
+                [86627] = { -- Incite (Cataclysm)
+                    [78]  = true, -- Heroic Strike
+                },
                 [52437] = { -- Sudden Death (Wrath+)
                     [5308] = true, -- Execute (not for Cata)
                     [86346]= true, -- Colossus Smash (Cataclysm)